dashboard/functions/requirements.txt
functions-framework==3.*

flask>=1.0.2

appengine-python-standard>=0.2.2
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-logging==3.1.2
grpcio
protobuf
